% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amirabin.R
\name{amira_app_path}
\alias{amira_app_path}
\alias{amira_version}
\alias{amira_start_path}
\title{Utility functions to locate Amira application}
\usage{
amira_app_path(path = NULL)

amira_version(path = getOption("nat.amira.amira"))

amira_start_path(path = NULL)
}
\arguments{
\item{path}{The path to the Amira application (defaults to the value of
\code{options('nat.amira.amira'))} or (on MacOS X) tries to find the
application.}
}
\description{
\code{amira_app_path} returns the path to the Amira application

\code{amira_start_path} returns the current version of Amira

\code{amira_start_path} returns the path to the Amira startup script
}
\examples{
\donttest{
amira_app_path()
amira_start_path()
amira_version()
}
}
