% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat.amira-package.r
\docType{package}
\name{nat.amira-package}
\alias{nat.amira}
\alias{nat.amira-package}
\title{A simple interface beween Amira and the neuronatomy toolbox (nat)}
\description{
The main function is \code{\link{open_amira}} which will open a variety of R
objects in the current running version of Amira. This works by writing a file
to disk in a suitable AmiraMesh file format and then opening it. The
following data types are supported:
}
\details{
\itemize{

\item points

\item neurons (as single objects)

\item \code{\link[nat]{neuronlist}} objects containing many neurons

\item surface objects

}

Package options:

#' \itemize{

\item nat.amira.amira The path to the Amira application (see
\code{\link{amira_app_path}} and \code{\link{open_amira}})

}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferis/nat.amira}
  \item Report bugs at \url{https://github.com/jefferis/nat.amira/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

}
\keyword{internal}
