% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat.amira-package.r
\docType{package}
\name{nat.amira}
\alias{nat.amira}
\alias{nat.amira-package}
\title{A simple interface beween Amira and the neuronatomy toolbox (nat)}
\description{
A simple interface beween Amira and the neuronatomy toolbox (nat)
}
