% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{write_amira_script}
\alias{write_amira_script}
\title{Write a simple Amira script}
\usage{
write_amira_script(lines, file = tempfile(fileext = ".hx"))
}
\arguments{
\item{lines}{Tcl code (not including Amira's header line)}

\item{file}{Output file (defaults to a temporary location). Should normally
have suffix \code{'.hx'}.}
}
\value{
The path to the output file (invisibly)
}
\description{
Write a simple Amira script
}
\examples{
\donttest{
open_amira(opensc<-write_amira_script("echo Hello!"))
}
}
