% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write_neurons_for_amira.R
\name{write_neurons_for_amira}
\alias{write_neurons_for_amira}
\title{write neurons out in a directory hierarchy suitable for Amira}
\usage{
write_neurons_for_amira(nl, rdir)
}
\arguments{
\item{nl}{a neuron list containing projection neurons}

\item{rdir}{The root directory of the project}
}
\description{
write neurons out in a directory hierarchy suitable for Amira
}

