% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/open.R
\name{open_amira}
\alias{open_amira}
\title{Open a file in Amira}
\usage{
open_amira(x = NULL, amira = getOption("nat.amira.amira", "Amira"))
}
\arguments{
\item{x}{A file to open (optional)}

\item{amira}{Path to desired Amira version (see details)}
}
\description{
Simply opens Amira if no file is specified. Presently only works on MacOS X.
}
\details{
Uses the package option otions('nat.amira.amira') to specify the path to
Amira when this is set.
}
\examples{
\dontrun{
## Open Amira
open_amira()

##
open_amira("myscript.scro")
open_amira("mynetwork.hx")
}
}

