% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_amira.default}
\alias{open_amira.default}
\alias{open_amira}
\alias{open_amira.neuron}
\alias{open_amira.neuronlist}
\alias{open_amira.matrix}
\alias{open_amira.data.frame}
\alias{open_amira.hxsurf}
\title{Open a file in Amira}
\usage{
\method{open_amira}{default}(x = NULL, ...,
  amira = getOption("nat.amira.amira", "Amira"))

open_amira(x = NULL, ...)

\method{open_amira}{neuron}(x, ...)

\method{open_amira}{neuronlist}(x, ...)

\method{open_amira}{matrix}(x, ...)

\method{open_amira}{data.frame}(x, ...)

\method{open_amira}{hxsurf}(x, ...)
}
\arguments{
\item{x}{A file to open (optional)}

\item{...}{Additional arguments passed to methods}

\item{amira}{Path to desired Amira version (see details)}
}
\description{
Simply opens Amira if no file is specified. Presently only works on MacOS X.

\code{open_amira.matrix} and \code{open_amira.data.frame} open a
  set of 3D points as Amira landmark

\code{open_amira.hxsurf} open a 3D surface mesh in Amira
}
\details{
Uses the package option otions('nat.amira.amira') to specify the path to
Amira when this is set.
}
\examples{
\dontrun{
## Open Amira
open_amira()

## Load some scripts
open_amira("myscript.scro")
open_amira("mynetwork.hx")

# Load some neurons and make checkboxes to turn on/off and colour by glomerulus
# nb ... is passed to write_neurons_for_amira in this case
open_amira(Cell07PNs, subdir=Glomerulus)
}
}
\seealso{
\code{\link{write_neurons_for_amira}}
}
