% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_stack_viewer}
\alias{open_stack_viewer}
\title{Open our Amira Stack viewer script to review a set of files}
\usage{
open_stack_viewer(stackdir = ".", filenames = NULL, template = c("JFRC2",
  "FCWB", "IS2", "T1"))
}
\arguments{
\item{stackdir}{Path to the directory containing stacks}

\item{filenames}{Character vector specifying restricted set of stacks to
view.}

\item{template}{Short name of template brain (to show surface)}
}
\description{
Open our Amira Stack viewer script to review a set of files
}
\examples{
\dontrun{
# vfbr library can download registered stacks from VFB
library(vfbr)
open_stack_viewer(getOption("vfbr.stack.downloads"))
}
}

